[#159] :recycle: refactor: 일기 배경, 일기 부분 공통 컴포넌트로 분리
