+++ client/src/components/common/diary/DiarySectionContainer.tsx
@@ -1,6 +1,6 @@
 import tw from "tailwind-styled-components";
 
 export const DiarySectionContainer = tw.div`
-bg-white/60 rounded-lg shadow-lg
-mx-auto my-[8vh] w-[90%] p-[3vh]
+relative bg-white/60 rounded-lg shadow-lg
+mx-auto w-[90%] p-[3vh] overflow-hidden
 `;
